+++ eval_utils.py
@@ -4,7 +4,7 @@ import torch
 import torchvision.transforms.functional as transforms_f
 from einops import rearrange
 
-from factorization_utils import factorize_labels
+from genie.factorization_utils import factorize_labels
 
 
 class AvgMetric:
